test: move compressor tests out of end2end_test.go
